requirements.txt
joblib==1.1.0
numpy==1.23.1
pandas==1.4.3
pendulum==2.1.2
prefect==2.0b10
PyYAML==6.0
scikit_learn==1.1.1
SQLAlchemy==1.4.39
xgboost==1.6.1
python-box==6.0.2
dvc==2.13.0